requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
html5lib==1.1
idna==3.2
lxml==4.6.3
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzlocal==3.0
urllib3==1.26.6
webbot==0.34
webdriver-manager==3.4.2
webencodings==0.5.1
